onedrive-mcp-server/pyproject.toml
[project]
name = "onedrive-mcp-server"
version = "0.1.0"
description = "MCP Server for OneDrive Integration"
requires-python = ">=3.11"

dependencies = [
    "mcp>=0.9.0",           # MCP Protocol
    "msgraph-sdk>=1.0.0",   # Microsoft Graph SDK
    "azure-identity>=1.14.0",  # Azure authentication
    "requests>=2.31.0",     # HTTP requests
    "python-dotenv>=1.0.0", # Environment variables
    "pydantic>=2.0.0",      # Data validation
    "httpx>=0.25.0",        # Async HTTP client
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]